requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
idna==3.6
Jinja2==3.1.3
langcodes==3.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.3
packaging==23.2
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.7.0/pl_core_news_sm-3.7.0-py3-none-any.whl#sha256=94fd294e49f87fccf863f2792a1d89be1a0fd0a3c549790b2c8338ac304bee36
preshed==3.0.9
progressbar2==4.3.2
pydantic==2.6.0
pydantic_core==2.16.1
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
python-utils==3.8.2
rapidfuzz==3.6.1
requests==2.31.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wasabi==1.1.2
weasel==0.3.4
